Tests updated

+++ api/views.py
@@ -1,5 +1,3 @@
-from django.http import JsonResponse
-from django.shortcuts import render
 from rest_framework import generics, status
 from .serializers import RoomSerializer, \
                          CreateRoomSerializer, \
@@ -46,13 +44,10 @@ class CreateRoomView(APIView):
     def post(self, request, format=None):
         if not request.user.is_authenticated:
             return Response({'Bad Request': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-        # if not self.request.session.exists(self.request.session.session_key):
-        #    self.request.session.create()
         serializer = self.serializer_class(data=request.data)
         if serializer.is_valid():
             guest_can_pause = serializer.data.get('guest_can_pause')
             votes_to_skip = serializer.data.get('votes_to_skip')
-            # host = self.request.session.session_key
             host = request.user
             queryset = Room.objects.filter(host=host)
             if queryset.exists():
@@ -63,7 +58,6 @@ class CreateRoomView(APIView):
             else:
                 room = Room(host=host, guest_can_pause=guest_can_pause, votes_to_skip=votes_to_skip)
                 room.save()
-            # self.request.session['room_code'] = room.code
             return Response(RoomSerializer(room).data, status=status.HTTP_201_CREATED)
         return Response({'Bad Request': 'Invalid data...'}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -74,47 +68,37 @@ class JoinRoom(APIView):
     def post(self, request, format=None):
         if not request.user.is_authenticated:
             return Response({'Bad Request': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-        # if not self.request.session.exists(self.request.session.session_key):
-        #    self.request.session.create()
         code = request.data.get(self.lookup_url_kwarg)
         if code != None:
             room_result = Room.objects.filter(code=code)
             if len(room_result) > 0:
                 room = room_result[0]
-                # self.request.session['room_code'] = code
                 return Response({'Message': "Room Joined"}, status=status.HTTP_200_OK)
             return Response({'Bad Request': "Invalid room code"}, status=status.HTTP_400_BAD_REQUEST)
         return Response({'Bad Request': "Invalid post data, did not find a code key"}, status=status.HTTP_400_BAD_REQUEST)
 
 
-class UserInRoom(APIView):
-    def get(self, request, format=None):
-        if not self.request.session.exists(self.request.session.session_key):
-            self.request.session.create()
-        data = {
-            'code': self.request.session.get('room_code')
-        }
-        return JsonResponse(data, status=status.HTTP_200_OK)
-
-
 class LeaveRoom(APIView):
+    lookup_url_kwarg = 'code'
     def post(self, request, format=None):
         if not request.user.is_authenticated:
             return Response({'Bad Request': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-        host_id = request.user
-        room_results = Room.objects.filter(host=host_id)
-        if len(room_results) > 0:
-            room = room_results[0]
-            room.delete()
-        return Response({'Message': 'Success'}, status=status.HTTP_200_OK)
+        code = request.data.get(self.lookup_url_kwarg)
+        if code != None:
+            room_results = Room.objects.filter(code=code)
+            if len(room_results) > 0:
+                room = room_results[0]
+                if room.host == request.user:
+                    room.delete()
+                return Response({'Message': 'Success'}, status=status.HTTP_200_OK)
+            return Response({'Bad Request': "Invalid room code"}, status=status.HTTP_400_BAD_REQUEST)
+        return Response({'Bad Request': "Invalid post data, did not find a code key"}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class UpdateRoom(APIView):
     serializerClass = UpdateRoomSerializer
 
     def patch(self, request, format=None):
-        # if not self.request.session.exists(self.request.session.session_key):
-        #    self.request.session.create()
         if not request.user.is_authenticated:
             return Response({'Bad Request': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
         serializer = self.serializerClass(data=request.data)
